// FOCUSSTARR/src/lib/types.ts
export interface Subtask {
  id: string;
  text: string;
  done: boolean;
}

export interface Task {
  id: string;
  title: string;
  project: string;
  priority: 'high' | 'medium' | 'low';
  status: 'todo' | 'doing' | 'review' | 'done';
  due: string;
  durationHours: number;
  durationMinutes: number;
  tags: string[];
  progress: number;
  notes: string;
  subtasks: Subtask[];
  isArchived: boolean;
}